refactor relay bridge layout and add python examples
