+++ internal/relay/adaptor/openai/response_bridge.go
